requirements.txt
Django==1.4
simplejson==2.2.1
lxml==2.3.5
Fabric==1.4.3
unittest2==0.5.1
HireFire==0.1
PyYAML==3.10
South==0.7.6
amqp==1.0.8
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.21
boto==2.5.2
celery==3.0.15
dj-database-url==0.2.1
django-appconf==0.6
django-bootstrap-toolkit==2.10.5
django-celery==3.0.11
django-profiles==0.2
django-ses==0.4.1
django-staticfiles==1.2.1
django-storages==1.1.5
django-tinymce==1.5.1
docutils==0.10
gevent==0.13.7
greenlet==0.4.0
gunicorn==0.14.6
kombu==2.5.6

mock==1.0b1
newrelic==1.11.0.55
psycopg2==2.4.5
pycrypto==2.6
python-dateutil==1.5
pytz==2012j
redis==2.7.1

six==1.3.0
ssh==1.8.0

wsgiref==0.1.2
#-e hg+https://bitbucket.org/ubernostrum/django-registration@b3c41b3c7f50523b4ecb8267d6d27a918adc2e1f#egg=django_registration-dev
